Fixed TemperatureStatic

+++ PrismWeatherApp.Core/TemperatureStatic.cs
@@ -5,8 +5,8 @@ namespace PrismWeatherApp.Core
     public static class TemperatureStatic
     {
         public static string CityName { get; set; }
-        public static double Latitiude { get; set; }
-        public static double Longitiude { get; set; }
+        public static float Latitiude { get; set; }
+        public static float Longitiude { get; set; }
         public static TemperatureHourly Hourly { get; set; }
     }
 }
